This is some early work on modules.
This is mostly a port of the jspp module work with some minor changes to the generated code.

This only supports inline modules (no require).

This contains a full port of the SemanticAnalyzer which contains logic for class and traits. Do you want me to remove that in this patch?

This also contains a port of codegenerator.Compiler and dependent classes. I need this because we need to provide the Project to maintain the semantic state.

Review URL: https://chromereviews.googleplex.com/2773013