This is a simple change to remove duplication between transformers and ensure traits are being transformed.
Basically what happen is: traits and classes used to have more sophisticated logic, and were split between TraitTransformer/ClassTransformer/AggregateTransformer. Now that it's so much simpler, everything is in ClassTransformer. (traits are treated as kind of a special, simple class). In the future, we'll need to fix ClassTransformer to walk the whole tree like most other program transformers, so this sets us up nicely for that.

Review URL: https://chromereviews.googleplex.com/2791013