Code/Module loaders
This implements load, eval and evalLoad. Right now these are not exposed to the code but the unit test creates the loader manually.

This does not implement:
- Globals
- Import
- Resolver

and the eval case is not doing eval in the right scope. Eval in general requires a lot of work even without modules.

BUG=
TEST=

Review URL: https://chromereviews.googleplex.com/2810022